Change possession when undoing startPoint
In normal game flow, possession changes from the scoring team to the new offense, so this accounts for that by calling changePossession in the memento for recordFirstActor. There are still some edge cases around the pull where multiple undo might result in the wrong possession, but this should fix it for the most common usage.